test: rewrite curp client tests
